move ToolCardLayout to a more appropriate source file
