+++ app/src/main/kotlin/org/cru/godtools/ui/tools/ToolCardLayouts.kt
@@ -1,17 +1,7 @@
 package org.cru.godtools.ui.tools
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.widthIn
 import androidx.compose.material3.CardDefaults.elevatedCardElevation
-import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.LocalTextStyle
 import androidx.compose.material3.MaterialTheme
@@ -22,7 +12,6 @@ import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
@@ -35,10 +24,8 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import coil.compose.AsyncImage
 import java.util.Locale
-import org.ccci.gto.android.common.androidx.compose.foundation.layout.widthIn
 import org.ccci.gto.android.common.androidx.compose.foundation.text.minLinesHeight
 import org.cru.godtools.base.ui.theme.GodToolsTheme
-import org.cru.godtools.base.ui.util.ProvideLayoutDirectionFromLocale
 import org.cru.godtools.base.ui.util.getCategory
 import org.cru.godtools.base.ui.util.withCompatFontFamilyFor
 import org.cru.godtools.model.getName
@@ -74,85 +61,6 @@ private val toolCardInfoLabelStyle @Composable get() = MaterialTheme.typography.
 sealed class ToolCardEvent(val tool: String?, val lang1: Locale? = null) {
     class Click(tool: String?, lang1: Locale? = null) : ToolCardEvent(tool, lang1)
     class OpenTool(tool: String?, lang1: Locale?) : ToolCardEvent(tool, lang1)
-}
-
-@Composable
-fun ToolCard(
-    state: ToolCard.State,
-    modifier: Modifier = Modifier,
-    confirmRemovalFromFavorites: Boolean = false,
-    showActions: Boolean = true,
-    interactionSource: MutableInteractionSource = remember { MutableInteractionSource() },
-) {
-    val translation by rememberUpdatedState(state.translation)
-    val secondLanguage by rememberUpdatedState(state.secondLanguage)
-    val downloadProgress by rememberUpdatedState(state.downloadProgress)
-    val eventSink by rememberUpdatedState(state.eventSink)
-
-    ProvideLayoutDirectionFromLocale(locale = { translation?.languageCode }) {
-        ElevatedCard(
-            onClick = { eventSink(ToolCard.Event.Click) },
-            elevation = toolCardElevation,
-            interactionSource = interactionSource,
-            modifier = modifier
-        ) {
-            Box(modifier = Modifier.fillMaxWidth()) {
-                ToolBanner(
-                    state,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .aspectRatio(335f / 87f)
-                )
-                FavoriteAction(
-                    state,
-                    confirmRemoval = confirmRemovalFromFavorites,
-                    modifier = Modifier.align(Alignment.TopEnd)
-                )
-                DownloadProgressIndicator(
-                    { downloadProgress },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .align(Alignment.BottomCenter)
-                )
-            }
-            Column(modifier = Modifier.padding(16.dp)) {
-                Row(
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    ToolName(
-                        state,
-                        modifier = Modifier
-                            .run { if (secondLanguage != null) widthIn(max = { it - 70.dp }) else this }
-                            .alignByBaseline()
-                    )
-                    if (secondLanguage != null) {
-                        ToolCardInfoContent {
-                            AvailableInLanguage(
-                                secondLanguage,
-                                horizontalArrangement = Arrangement.End,
-                                modifier = Modifier
-                                    .padding(start = 8.dp)
-                                    .alignByBaseline()
-                            )
-                        }
-                    }
-                }
-                ToolCategory(state, modifier = Modifier.fillMaxWidth())
-
-                if (showActions) {
-                    ToolCardActions(
-                        state,
-                        buttonWeightFill = false,
-                        buttonModifier = Modifier.widthIn(min = 92.dp),
-                        modifier = Modifier
-                            .padding(top = 4.dp)
-                            .align(Alignment.End)
-                    )
-                }
-            }
-        }
-    }
 }
 
 @Composable
