+++ app/src/test/snapshots/images/org.cru.godtools.ui.tools_ToolCardPaparazziTest_ToolCard()_-_Default[NIGHT,NO_ACCESSIBILITY].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:1209e7696964600e1ed4f24d35d3a3eb3033c02611c50318ae23faa2c07eeb6a
+size 43027
