+++ app/src/test/snapshots/images/org.cru.godtools.ui.tools_ToolCardPaparazziTest_ToolCard()_-_Default[NOTNIGHT,ACCESSIBILITY].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:876468ad27d35dafa8015e3df651b6a04bc25ea0f95cae38500d16b7f0d111f2
+size 52437
