+++ app/src/test/snapshots/images/org.cru.godtools.ui.tools_ToolCardPaparazziTest_ToolCard()_-_Default[NOTNIGHT,NO_ACCESSIBILITY].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:8a819793c463a89544905a8e4da3e400280feda71fa2cab1296b99299967e8d5
+size 43143
